fix(lint): TASK-008 — repay clippy baseline drift, 10 minimal mechanical sites
Warn counts had drifted 349 -> 354 actual across 28 commits since the
WO-24 lock, so the WO-23 warn-baseline gate was red on HEAD (would
fail CI clippy on push). Ten sites fixed mechanically, no logic
changes:
- same-type cast removals x6 (image_base/int_val already u64)
- checked_add().unwrap_or(u64::MAX) -> saturating_add()
- map_or(false,..) -> is_some_and(..) x2
- BTreeMap tuple -> type PreTruncParentRow alias
- unit let-bindings removed x4, unused param/unsafe cleaned,
  digit grouping unified
Baseline lowered in the same commit (only downward, per policy):
TOTAL 349 -> 337.

Acceptance verified personally by the commander, all three green:
baseline gate exit 0 (TOTAL=337) / cargo test --workspace --offline
2801 passed 0 failed / cargo fmt exit 0. Skipped sites documented
(&dyn Fn aliases need lifetime params, E0373, beyond minimal fix).
Archive: runs/20260829-TASK-008.md.

+++ crates/cli/src/unpacker/iat_observe.rs
@@ -302,7 +302,7 @@ pub fn iat_observe(
                 // hint/name RVA: read the name at int_val (the IMAGE_IMPORT_BY_NAME
                 // struct starts with a hint u16 then the name).
                 let mut hdr = [0u8; 2];
-                let _ = read_at(image_base + int_val as u64, &mut hdr);
+                let _ = read_at(image_base + int_val, &mut hdr);
                 let name_rva_import = int_val as u32 + 2;
                 read_cstring_at(&read_at, image_base, name_rva_import, 128).unwrap_or_default()
             };
